Main code
this is the main (original) code of the converter. it perfectly converts german language to the "löffelsprache". 
Just run it in a terminal or console window with >> .\ python3 spoonish.py (without any parameters) <<
you will be promted to input a text which will be converted after you press Enter. 
If you feel like it you can remove the input from the code and modify the function, 
so you can call the script directly with the text as a parameter - you will need to add a return in this case.